refactor(trusty-review): split the --analyze lane out of run.rs (Refs #6811)
#6811 pushed run.rs to 525 SLOC, over the 500 production cap. The seam is
the one the pipeline already had: run_analyze.rs holds everything about ONE
optional lane — its per-request budget, its fetch, and whether its outcome
may ship — while run.rs keeps orchestrating the run as a whole. Same
principle that carved analyze_enrich.rs off analyze_adapter.rs. The four
functions move verbatim apart from their imports and pub(super) visibility.
