+++ crates/trusty-review/src/report/run.rs
@@ -25,6 +25,7 @@ use crate::report::investigate::{
 };
 use crate::report::manifest::Manifest;
 use crate::report::model::{InferenceAttribution, ReportModel, RoleAttribution};
+use crate::report::run_analyze::{analyze_lane_verdict, enrich_from_analyze};
 use crate::report::synthesize::{Synthesis, Synthesizer, ground_investigation_prose};
 use crate::report::template::{DEFAULT_TEMPLATE, TemplateLoader, resolve_template_alias};
 use crate::report::{
@@ -418,159 +419,6 @@ fn flag_is_truthy(raw: &str) -> bool {
     matches!(
         raw.trim().to_ascii_lowercase().as_str(),
         "1" | "true" | "yes" | "on"
-    )
-}
-
-/// The per-request budget the corpus-scanning analyze endpoints get (#6712).
-///
-/// Why: a free function so the precedence is tested directly rather than
-/// restated inside the fetch, exactly as `resolve_code_only_from` is.
-/// What: the request's value, else the manifest's, else the default; a `0` at
-/// either tier reads as absent, since zero would time out every request.
-/// Test: `run_tests::the_request_analyze_timeout_beats_the_manifest`.
-fn resolve_analyze_budget(req: &ReportRequest, manifest: &Manifest) -> std::time::Duration {
-    crate::report::analyze_endpoints::corpus_budget_from_secs(
-        req.analyze_timeout_secs
-            .or(manifest.report.analyze_timeout_secs),
-    )
-}
-
-/// Run the opt-in deterministic analyze fetch, recording what it could not
-/// reach (#2445, #5239).
-///
-/// #6712: the corpus-scanning endpoints take their per-request budget from
-/// `--analyze-timeout-secs`, else the manifest's `[report].analyze_timeout_secs`,
-/// else the default — the same request-beats-manifest precedence `--code-only`
-/// and `--no-mermaid` already follow.
-/// Test: `run_tests::the_request_analyze_timeout_beats_the_manifest`.
-async fn enrich_from_analyze(
-    model: &mut ReportModel,
-    req: &ReportRequest,
-    manifest: &Manifest,
-    config: &ReviewConfig,
-) -> crate::report::AnalyzeLaneCoverage {
-    let analyze_socket = manifest
-        .report
-        .analyze_socket
-        .clone()
-        .map_or_else(|| config.analyzer_socket.clone(), std::path::PathBuf::from);
-    let corpus_budget = resolve_analyze_budget(req, manifest);
-    eprintln!(
-        "[trusty-review report] --analyze: fetching over {} (corpus endpoints allowed {corpus_budget:?} each)",
-        analyze_socket.display()
-    );
-    match crate::report::HttpAnalyzeMetricsSource::new(analyze_socket) {
-        Ok(source) => {
-            let source = source.with_corpus_budget(corpus_budget);
-            // #5239: every repo the fetch could not populate is named in the
-            // report, not only warned about on stderr — a dimension missing
-            // because the daemon was down must not read as a clean pass.
-            let outcome = crate::report::enrich_with_analyze_outcome(model, &source).await;
-            for gap in &outcome.lines {
-                eprintln!("[trusty-review report] --analyze gap: {gap}");
-            }
-            model.gaps.extend(outcome.lines);
-            outcome.coverage
-        }
-        Err(e) => {
-            eprintln!(
-                "[trusty-review report] --analyze: could not build HTTP client ({e}); \
-                 falling back to scan"
-            );
-            // #5239: the client never existed, so no repo was assessed — that is
-            // a whole-report gap, not a per-repo one.
-            model.gaps.push(
-                "trusty-analyze data unavailable — the analysis client could not be built, so no \
-                 application in this report was assessed against trusty-analyze. Findings, \
-                 complexity, and health factors are not assessed, not clean."
-                    .to_string(),
-            );
-            // #6811: no client means no repository was assessed, which is the
-            // same total collapse the walk reports — counted over the
-            // repositories the walk WOULD have attempted, so the verdict below
-            // sees one condition rather than two.
-            crate::report::AnalyzeLaneCoverage::never_ran(analyze_eligible_repositories(model))
-        }
-    }
-}
-
-/// Repositories the analyze walk would attempt, counted without walking (#6811).
-///
-/// Why: the client-construction failure arm never reaches
-/// `enrich_with_analyze_outcome`, so it has no denominator of its own — and a
-/// total collapse with no denominator cannot be told from a manifest that had
-/// nothing to assess. This applies the walk's own eligibility rule minus the
-/// index-registry lookup a dead client cannot perform, so it can only
-/// over-count (a checkout whose path derives to no index id), never under-count.
-/// What: local checkouts that declared no metrics file.
-/// Test: `run_tests::a_client_that_will_not_build_counts_every_eligible_repository`.
-fn analyze_eligible_repositories(model: &ReportModel) -> usize {
-    model
-        .repositories
-        .iter()
-        .filter(|repo| repo.metrics.is_none() && repo.local_path.is_some())
-        .count()
-}
-
-/// Whether an analyze lane's outcome may ship (#6811).
-///
-/// Why: DOC-67 §9's fail-open contract is per repository and stays that way.
-/// What it never covered is the lane failing for EVERY repository, which #6783
-/// shipped: 59 repositories all reading "not stated in source data", which a
-/// downstream reader took for "static analysis ran and found nothing". A report
-/// nothing was measured for is not a degraded report, so the run fails unless
-/// the operator says otherwise.
-///
-/// The threshold: only a TOTAL collapse (`0 of N` assessed, `N > 0`) fails.
-/// Partial degradation (`M of N`, `M > 0`) stays a warning at every setting — a
-/// 58-of-59 run carries 58 assessed applications, and throwing that away over
-/// one unreachable index costs more than it protects. A lane attempted zero
-/// times is not a failure either: a manifest of remote-only repositories has no
-/// analyze lane to lose.
-///
-/// # Errors
-///
-/// When the lane assessed nothing and `allow_degraded` is false. The message
-/// names the lane, both counts, and the flag that overrides it.
-///
-/// Test: `run_tests::{a_total_analyze_collapse_fails_the_run,
-/// allow_degraded_lets_a_total_collapse_ship,
-/// a_partial_analyze_degradation_is_a_warning_not_a_failure,
-/// an_unattempted_analyze_lane_is_not_a_failure}`.
-fn analyze_lane_verdict(
-    coverage: crate::report::AnalyzeLaneCoverage,
-    allow_degraded: bool,
-) -> Result<()> {
-    if !coverage.is_total_failure() {
-        if coverage.is_partial_failure() {
-            eprintln!(
-                "[trusty-review report] --analyze: lane partially degraded — {} of {} \
-                 application(s) assessed, {} failed; the report states the coverage and the run \
-                 continues (issue #6811)",
-                coverage.succeeded,
-                coverage.attempted,
-                coverage.failed(),
-            );
-        }
-        return Ok(());
-    }
-    if allow_degraded {
-        eprintln!(
-            "[trusty-review report] --analyze: lane DID NOT RUN — 0 of {} application(s) \
-             assessed; --allow-degraded was passed, so the report is written with that stated \
-             (issue #6811)",
-            coverage.attempted,
-        );
-        return Ok(());
-    }
-    anyhow::bail!(
-        "the trusty-analyze lane DID NOT RUN — 0 of {attempted} application(s) assessed, \
-         {failed} failed. Every finding count, complexity figure and health factor this report \
-         would carry is UNASSESSED, not clean, so it is not written. Fix the analyze lane (is \
-         `trusty-analyze` running, and is each checkout indexed?) and re-run, or pass \
-         --allow-degraded to ship the report with that fact stated in it.",
-        attempted = coverage.attempted,
-        failed = coverage.failed(),
     )
 }
